change --insert_code_prefetches to specify number of PRFM instructions
make --insert_code_prefetches take an integer parameter for the number
of PRFM instructions to use for each target function.   Also, generate
a series of NOPs that match the number of PRFM instructions so that if
ENABLE_CODE_PREFETCH is not defined, the function sizes are the same.

Use __asm__ __volatile__ for inserting prefetch instructions.